Bump to latest gitlab patch version
This also gets rid of the custom images we were using to set our own
needs limit and yaml size limit.  The needs limit setting was no
longer valid anyway after [1], and the yaml size limit is not
likely to be required anymore either, now that we pass one spack.lock
to downstream jobs via artifacts rather than encoding all the root
specs as variables in the yaml and duplicating them many times.  The
largest generated pipeline yaml I found recently was when we turned
back on all of E4S, and in that case the yaml was just over 500 KB,
and the default limit is 1MB.

Notes on needs limits:

"ci_needs_size_limit" seems to be the value used now to determine the
limit as seen the ruby file of the current webservice image:

    def max_needs_allowed
      @pipeline.project.actual_limits.ci_needs_size_limit
    end

I used the gitlab rails console to look for the current needs limit and
it seems if it's enabled at all in our instance, it's set to 100:

    irb(main):004:0> Plan.default.actual_limits
    => #<PlanLimits id: 1,...,ci_needs_size_limit: 100,...>
    irb(main):005:0>

[1] https://gitlab.com/gitlab-org/gitlab/-/merge_requests/37568