Fix yaml size limits issue
Even though it appears we should no longer need this, we are hitting
yaml size limits when pipeline generation creates a yaml file of
around 400 KB on disk.  This change replaces the patching of docker
images we previously used to get around the yaml size limitations,
while keeping us on the latest patch release of gitlab.