Support spackbot fixing style in worker process
This PR adds a worker pod to the spackbot deployment which will handle
asynchronously the work of fixing isort style for PRs.  The worker pod
can be used for other asynchronous tasks in the future as well (e.g. PR
binary graduation), and we can increase the number of replicas if
necessary to handle the extra load.