Fix spackbotdev workflow documentation
Update documentation to reflect reality as learned recently while
practicing the development workflow for spackbot.  Adding annotations
on the pods themselves doesn't prevent flux from noticing your test
changes to the deployments and "fixing" them for you.  Instead,
the annotation telling flux to ignore changes needs to be applied
to the deployments.