Start specifying resource requests
Most of the things we deployed in the cluster do not specify the
amount of memory or cpu they request, this information is critical
to allow kubernetes to make good pod scheduling decisions.  This
PR applies cpu and memory requests to some of the large resource
consumers that did not previously specify resource requests.