Reduce spackbot-workers memory request
Autoscaler is having trouble scheduling this pod because of memory
issues.  The t3.small have 2G ram, so you would think we could safely
request 1.3G.  And in fact, it did get scheduled and run for a little
while with the request set to 1.3G, but was still found pending an
hour later.