gh-gl-sync: Make sure merge commit message matches commit parents
We saw recently that fetching from the repo can produce a different
PR HEAD sha than the GH api.  When that happened, we used the value
from the api (pull.head.ref) to make the merge commit message, but
the commit contained an older PR HEAD as one of its parent.

This change checks to make sure the two values match.  If they do
not match, since we cannot be sure which sha is correct, we avoid
making and pushing the merge commit for the PR.