Wait an hour before considering a pipeline as skipped
Update our skipped pipelines mitigation to wait an hour before
using the API to trigger a pipeline for a branch that otherwise
does not have one for its HEAD commit.

This resolves a problem we noticed recently where GitLab eventually
reacted to the original push event. That resulted in two pipelines
for a given commit.