Add gitlab OIDC provider
This adds a gitlab OIDC provider to the spack module.

It also creates multiple resources for both the protected binaries
bucket and the prs binaries bucket*. The roles it creates can be
assumed only when the JWT matches the claims in
matching_subjects. This means that when a request for
protected_mirrors comes in to the provider, it must also match the sub
claim configured here.

This is the most concise explanation of how the overall flow works: https://docs.gitlab.com/ee/ci/cloud_services/index.html#authorization-workflow.

*The production bucket policies/roles aren't yet modularized in
Terraform, so there's some logic to use the existing ones for
production and create new ones for staging.