Add timestamp checking on archives
There was one issue in this script that would only check timestamps on
the pipeline but not the archive, which lead to race conditions when
pruning using a time stamp provided by the other pruning method.

Now the timestamp cutoff for "updated-before" is respected by both the
get_pipelines and generate_pruning_lists methods.