Publish stacks within a ref to the top-level (#848)
Add a container image with tooling to support publishing binary mirrors from the
stack-specific mirrors in which pipelines build them, to a top-level mirror, for easier
consumption by end users. See the README.md for complete details.  Also add
a cronjob to run the publish script every three hours, and add new iam role and
policy to give permission to write to the protected binary mirror.

---------

Co-authored-by: Zack Galbreath <zack.galbreath@kitware.com>