Only store concrete packages in package dimension
Prior to this, the package dimension was used very loosely. Now, it stores a concrete spec,
and only the 'variants' field can be empty (since that field may be empty normally).

To achieve this, the build timing code is updated to fetch the spec json from gitlab and read in all the concrete specs.