Enable stesachs@amazon.com to run `s3 sync` on a sub-bucket
I want to add to the existing buildcache location and run `spack buildcache
update-index` in order to merge the content. Spack needs to be able to list the
content (or use `aws s3 sync`) to run `buildcache update-index`.

This is an extension of #730 to enable `aws s3 sync` additionally to `aws s3 cp`.