Fix analytics webhook crash on non-retryable jobs
We've been seeing this crash for a while now, where the job retry
reasons field doesn't exist. This is actually expected behavior, as
there is now a subset of jobs that we do not retry (no-spec-to-rebuild
jobs; see https://github.com/spack/spack/pull/46713/commits/e988ce058133f74066709ae8217bdb9add12190d).
This went into effect about a month ago, which lines up rougly with when
we first started seeing this crash.