Remove service tag from graviton2-pub runners
We continue to see a lot of pod_failed errors for this runner.
The fact that it bears the service tag is a primary difference between
it and the graviton3-pub runners, which do not suffer from this same
error rate.

Try removing the service tag to see if these runners start performing
better.

Note that the x86-64-v2-pub runner still bears the "service" tag, and
it is not used for many other types of jobs.