cache_indexer: Add support for newest spack buildcache layout v3
Iterates bucket looking for index manifests, rather than the indices
themselves.  Each manifest is retrieved in order to get the content
address of the index, then the url to the index content blob is stored
under the "url" key.

Support for v3 is still hidden behind a comment, so one more PR to
switch that commenting is required to actualy change from v2 to v3.