Enable `serializeImagePulls` for karpenter linux nodes
The Amazon Linux 2023 AMI overrides the default kubelet config to
disable `serializeImagePulls` in order to improve performance. This is
not ideal for our use case, where we frequently create many pods at once
when start a CI pipeline, because it can cause us to hit the container
registry's max QPS, so we override it here.