Add `github_down` error
This is starting to happen because we are now cloning the spack repo
from github inside CI jobs (as opposed to before when we used
`spack/spack`, because `spack` was already available).
The GitLab regex was not specific enough, causing jobs like the below
example to be incorrectly marked as `gitlab_down`.

Example:
https://gitlab.spack.io/spack/spack-packages/-/jobs/18413048